Convert res to a long long int variable
sizeof (long int) is equal to sizeof(int) on x86 architecture, so res cannot
detect whether str is overflow or not. We must use a long long int variable.

+++ src/utils.c
@@ -136,7 +136,7 @@ int str2unshort(char *str, unsigned short *num)
 
 int strtoint(char *str, int *num)
 {
-	long int    res = 0;
+	long long int    res = 0;
 	int     len;
 	int     i;
